Extract `showGlobalErrorMessage` to `modules/message.ts`
Move the function out of bootstrap.ts into a side-effect-free module so
the ESM bundle can import it without re-triggering the global error
handler initialization. Fix overflow-menu popup item styling to override
fomantic specificity.
